San Juan: RPi 5: rpi-firmware: Add support for multiple autoboot
As discussed in
https://github.com/rauc/rauc/pull/1599#discussion_r1919769970,
swapping two versions of autoboot.txt is a better option than
modifying autoboot.txt.
My previous modifications to the rpi-firmware package assumed
that there would only be one autoboot.txt, but we now need to
also include the configuration for the slots, which will be
swapped in place of autoboot.txt upon update.

Add an option in rpi-firmware to put additional autoboot files
in the boot partition.

Signed-off-by: Olivier Benjamin <olivier.benjamin@bootlin.com>