San Juan: RPi 5: rpi-firmware: Add support for multiple autoboot
As discussed in
https://github.com/rauc/rauc/pull/1599#discussion_r1919769970,
swapping two versions of autoboot.txt is a better option than
modifying autoboot.txt.
My previous modifications to the rpi-firmware package assumed
that there would only be one autoboot.txt, but we now need two.

Add an option in rpi-firmware to put two autoboot.txt files in
the boot partition, as autoboot.txt and autoboot_alt.txt.

Signed-off-by: Olivier Benjamin <olivier.benjamin@bootlin.com>